add missing pkg and update README in hobbes-nvl-bb
Adding the ca-certificate to the hobbes-nvl-bb that
appears to be required for git clones from github.
This should be in the base image but adding here
to be double sure.

Update README in Hobbes commands area to to show very
basic launch into docker container and running qemu with
VM image.